renaming converter component to Rgb-to-Hex Component
