+++ src/Components/Home/Home.tsx
@@ -37,7 +37,7 @@ const Home = () => {
                <div className="card-body p-5">
                   <h5 className="card-title">Explore App Features</h5>
                   <p className="card-text">Something More About this App Let's Go...</p>
-                  <a href={ROUTES.CONVERTER} className="btn btn-primary">Go to Converter</a>
+                  <a href={ROUTES.CONVERTER} className="btn btn-primary">Go to App</a>
                </div>
             </div>
          </section>
